Flatten NodeDiagnostics interface
