+++ src/freenet/node/diagnostics/ThreadDiagnostics.java
@@ -1,7 +1,7 @@
 /* This code is part of Freenet. It is distributed under the GNU General
  * Public License, version 2 (or at your option any later version). See
  * http://www.gnu.org/ for further details of the GPL. */
-package freenet.node.diagnostics.diagnostics;
+package freenet.node.diagnostics;
 
 public interface Diagnostics {
     void start();
